#22 Move files
Organise files based on routes.
