+++ src/app/app-routing.module.ts
@@ -3,9 +3,9 @@ import { NgModule } from '@angular/core';
 import type { Routes } from '@angular/router';
 import { RouterModule } from '@angular/router';
 import { LoginComponent } from './login/login.component';
-import { ProgramComponent } from './program/program.component';
-import { AddEventComponent } from './add-event/add-event.component';
-import { EditEventComponent } from './edit-event/edit-event.component';
+import { ProgramComponent } from './program/view/program.component';
+import { AddEventComponent } from './program/add/add-event/add-event.component';
+import { EditEventComponent } from './program/edit/edit-event/edit-event.component';
 
 //insert here more routes when needed
 const routes: Routes = [
