+++ src/app/app.module.ts
@@ -11,25 +11,25 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { LoginComponent } from './login/login.component';
-import { ProgramComponent } from './program/program.component';
+import { ProgramComponent } from './program/view/program.component';
 import { DummyDialogComponent } from './dummy-dialog/dummy-dialog.component';
 import { RouterLink } from '@angular/router';
-import { AddEventComponent } from './add-event/add-event.component';
+import { AddEventComponent } from './program/add/add-event/add-event.component';
 import { MatSelectModule } from '@angular/material/select';
-import { ProgramFiltersComponent } from './program/program-filters/program-filters.component';
+import { ProgramFiltersComponent } from './program/view/program-filters/program-filters.component';
 import { MatLegacyChipsModule } from '@angular/material/legacy-chips';
 import { MatChipsModule } from '@angular/material/chips';
 import { NavigationComponent } from './navigation/navigation.component';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
-import { EventFormComponent } from './event-form/event-form.component';
-import { DeleteDialogComponent } from './program/delete-dialog/delete-dialog.component';
-import { EventFormEditButtonsComponent } from './event-form/event-form-edit-buttons/event-form-edit-buttons.component';
-import { ProgramViewComponent } from './program/program-view/program-view.component';
+import { EventFormComponent } from './program/shared/event-form/event-form.component';
+import { DeleteDialogComponent } from './program/view/delete-dialog/delete-dialog.component';
+import { EventFormEditButtonsComponent } from './program/shared/event-form/event-form-edit-buttons/event-form-edit-buttons.component';
+import { ProgramViewComponent } from './program/view/program-view/program-view.component';
 import { MatCardModule } from '@angular/material/card';
-import { EditEventComponent } from './edit-event/edit-event.component';
-import { CategoryDialogComponent } from './event-dialog/category-dialog/category-dialog.component';
-import { EventDialogComponent } from './event-dialog/event-dialog.component';
+import { EditEventComponent } from './program/edit/edit-event/edit-event.component';
+import { CategoryDialogComponent } from './program/shared/event-dialog/category-dialog/category-dialog.component';
+import { EventDialogComponent } from './program/shared/event-dialog/event-dialog.component';
 import {
   NgxMatDatetimePickerModule,
   NgxMatNativeDateModule,
