+++ src/app/program/shared/event-dialog/category-dialog/category-dialog.component.spec.ts
