+++ src/app/program/shared/event-dialog/event-dialog.component.ts
