+++ src/app/program/view/delete-dialog/delete-dialog.component.ts
